requirements_rocm.txt
matplotlib==3.9.4
pillow==11.1.0
